Streamline search building
