+++ package.json
@@ -5,10 +5,10 @@
 	"private": true,
 	"scripts": {
 		"dev": "next dev --turbopack",
-		"build": "next build && node generate-routes.js && PUPPETEER_ARGS='--no-sandbox --disable-setuid-sandbox' && node build-search.js",
+		"build": "next build && node scripts/build-search.js",
 		"build:without-search": "next build",
 		"start": "npx serve@latest out",
-		"build-search": "node generate-routes.js && PUPPETEER_ARGS='--no-sandbox --disable-setuid-sandbox' && node build-search.js",
+		"build-search": "node scripts/build-search.js",
 		"format:pages": "prettier --write 'app/**/*.{js,ts,tsx,mdx}'",
 		"generate-seid-docs": "./generate-seid-docs.sh",
 		"check-links": "node scripts/check-links.mjs"
