+++ scripts/check-links.mjs
@@ -1,4 +1,3 @@
-// scripts/check-pages.mjs
 import fs from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
